dom/imported-tests/webapps/WebStorage/tests/submissions/Infraware/test_storage_session_removeitem.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>storage_session_removeItem</h1>
    <div id="log"></div>
    <script>
        test(function() {
            sessionStorage.clear();
            sessionStorage.setItem("name", "user1");
            assert_equals(sessionStorage.getItem("name"), "user1");

            sessionStorage.removeItem("name");
            sessionStorage.removeItem("unknown");
            assert_equals(sessionStorage.getItem("name"), null, "sessionStorage.getItem('name')")
        });
    </script>
 </body>
</html>
